Small refactoring - use shorter syntax - use isolated С realisation - base for create(link)
